Tests/LibWeb/Text/input/HTML/document-open-during-pageshow-stops-stale-completion.html

<!DOCTYPE html>
<script src="../include.js"></script>
<script>
    asyncTest(async done => {
        const iframe = document.createElement("iframe");
        const initialLoad = new Promise(resolve => iframe.addEventListener("load", resolve, { once: true }));
        document.body.appendChild(iframe);
        await initialLoad;

        const iframeDocument = iframe.contentDocument;
        let iframeLoadCount = 0;
        iframe.addEventListener("load", () => ++iframeLoadCount);
        iframeDocument.open().write("<body>First document</body>");

        iframe.contentWindow.addEventListener("pageshow", () => {
            iframeDocument.open().write("<body>Replacement document</body>");
            println("Replaced parser during pageshow event");
            iframeDocument.addEventListener("readystatechange", () => {
                if (iframeDocument.readyState !== "complete")
                    return;
                println("Replacement reached complete ready state");
                println(`Stale iframe load events: ${iframeLoadCount}`);
                iframe.remove();
                done();
            });
            iframeDocument.close();
        }, { once: true });

        iframeDocument.close();
    });
</script>
