Tests/LibWeb/Text/expected/HTML/document-open-during-pageshow-stops-stale-completion.txt

Replaced parser during pageshow event
Replacement reached complete ready state
Stale iframe load events: 0
